Rename `duration` module to `time_delta`
